+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to Fluentify will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2024-12-12
+
+## Changed
+
+- Updated Microsoft.CodeAnalysis.CSharp to Version 4.12.0 (#26).
+- Updated Microsoft.CodeAnalysis.CSharp.Workspaces Version 4.12.0 (#26).
+- Updated Microsoft.CodeAnalysis.Workspaces.Common Version 4.12.0 (#26).
+
 ## [1.3.0] - 2024-12-03
 
 ## Added